Add files via upload
Program Description: This main program (LetterInventory.java) implements ArrayLists to keep track of an inventory of letters in the alphabet. It takes in a String, computes the number of each letter in the alphabet, and returns an inventory of the respective counts. Any non-alphabetic characters along with the casing of all letters are ignored.